refactor(#447): split natureSoundSynth.ts into natureSounds/helpers + natureSounds/sounds
